add basket & product

+++ javaAdvanced-21/src/main/java/hw/dao/ProductRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import hw.domain.Product;
 
-public interface ProductRepo extends JpaRepository<Product, Long>{
+public interface ProductRepo extends JpaRepository<Product, Integer>{
 
 }
